% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eigvalexample}
\alias{eigvalexample}
\title{Eigenvalues of PCA}
\format{
numeric vector
}
\source{
Simulated data
}
\usage{
data(eigvalexample)
}
\description{
A numeric vector containing the eigenvalues of the PCA analysis.
}
\keyword{datasets}
