% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kann.R
\name{kann}
\alias{kann}
\title{KANN Algorithm}
\usage{
kann(
  ref.profiles = NULL,
  dist.matrix = NULL,
  k = 25,
  p = 1,
  r0 = 0.1,
  cores = 1
)
}
\arguments{
\item{ref.profiles}{data.table of reference sample profiles (required)}

\item{dist.matrix}{matrix of pairwise distances between query (row) and reference (column) samples (required)}

\item{k}{Number of nearest samples (positive integer) to consider in k-nearest neighbor regression}

\item{p}{Power of the inverse distance in which to raise reference sample weights}

\item{r0}{Minimum radius inside which samples are assigned same weight r0}

\item{cores}{Number of parallel CPU cores utilized}
}
\value{
data.table of query sample profiles
}
\description{
Estimates ancestry fractions for a set of query individuals based on
distances in principal components space and reference sample ancestries.
The function requires the distance matrix as an input, see documentation for
distance_matrix function.
}
\examples{

data(pcaexample)
data(eigvalexample)
data(refidexample)
data(refexample)
data(queryidexample)

dist <- distance_matrix(eigenvec = pcaexample,
                        eigenval = eigvalexample,
                        id.ref = refidexample,
                        id.query = queryidexample,
                        pcs = 2)

query.profiles <- kann(ref.profiles = refexample,
                       dist.matrix = dist,
                       k = 25,
                       p = 0,
                       r0 = 0.1,
                       cores = 1)

}
