% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pcaexample}
\alias{pcaexample}
\title{Example PCA eigenvector data}
\format{
data table:
\describe{
\item{ID}{character, sample identifier (e.g., "Ref1", "Query1")}
\item{PC1}{numeric, 1st principal component coordinate}
\item{PC2}{numeric, 2nd principal component coordinate}
}
}
\source{
Simulated data
}
\usage{
pcaexample
}
\description{
A data set containing PCA coordinates and IDs for reference and query samples.
}
\keyword{datasets}
