% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{queryidexample}
\alias{queryidexample}
\title{Query sample IDs}
\format{
character vector
}
\source{
Simulated data
}
\usage{
queryidexample
}
\description{
A character vector with IDs for query individuals.
}
\keyword{datasets}
