% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{refidexample}
\alias{refidexample}
\title{Reference sample IDs}
\format{
character vector
}
\source{
Simulated data
}
\usage{
refidexample
}
\description{
A character vector with IDs for reference individuals.
}
\keyword{datasets}
